Most notably, added a new class "FPImage" which provides a wrapper around some pyfits utilities and some methods I use often. Have implemented this everywhere throughout the code and it seems to work well. I think it's made a lot of the code much more readable.
Made most of the code (mostly) PEP 8 compliant. There's a few
82-character long lines here and there and my docstrings are more than
72 characters, but I feel that changing them makes the code less
readable.

Fixed a major bug in the velocity map fitting routine where the
variances (and not the uncertainties) were being fed to the line fitter.
This was causing absurdly large resulting uncertainties in the fitted
line profiles.

Eliminated "fit_sky_level". This is now a method in the new FPImage
class.

Eliminated "identify_objects". This is irrelevant since the pipeline now
uses photutils instead of IRAF for photometry.

Eliminated "make_single_extension". Single extension fits files have
their advantages, but they are far outweighed by the convenience of
multi-extension fits files. IRAF's geotran is the last holdout for
things that need single-extension files in this code, and there's a
silly workaround that seems to be doing the job correctly.