Two major changes to the code in this update:
(1) Fixed a number of things related to the propagation of uncertainties
through the pipeline. A combination of bugs in the image convolution
code and image shifting/normalization code have been fixed. Note that
the image alignment now happens at the same time as the convolution step
at the end, rather than its old location. Uncertainties should now be
correct as far as I can tell. Uncertainty estimates from earlier
versions of this code should not be trusted.

(2) Added a manual ring profile fitting option for the sky subtraction
step, which can be triggered by using the "X" key. This mode is
cumbersome and more of an art than a science, so I'd recommend using it
only if you're having a LOT of trouble fitting a particular image.

Minor change: sky ring subtraction now saves after *every* image rather
than saving all images at the end of the process. This makes the code
much easier to run in small bits of time, since this was one of the most
time-consuming parts of the pipeline and intermediate saves are usually
a good idea.

Minor change: Velocity map fitting code now points to a specific
location for the "voigtfit.so" module, so you can have that module in
one specific location on your computer rather than in every directory in
which you wish to run the code. Make sure to point the code to the right
directory on your machine when you install it by editing the import line
in "fit_velmap.py"