mask_regions.py now much more pythonic and modular
Previously, mask_regions.py read in a region file, and called a series
of different functions (depending on the object shape) which all
returned the masked array (with 1 being the only possible mask value).

Now, mask_regions has a function "parse_region_file()" which returns the
parsed region information, and this information can then be fed to
various "mask_{shape}()" functions. The masking functions now return a
boolean mask array the same size as the input array, rather than the
masked array itself. In this way, the outputs can be used more generally
in other code.

The functionality of mask_regions() itself has not changed; only the
internal structure of the other functions for easier use outside the
pipeline.